+++ seafarers_of_splorr/State.InPlay.IslandBuy.cpp
@@ -38,7 +38,7 @@ namespace state::in_play::IslandBuy
 	}
 
 	static std::map<game::Item, double> unitPrices;
-	static int hiliteRow = 0;
+	static size_t hiliteRow = 0;
 
 	static void UpdateUnitPrices()
 	{
